last

+++ Assets/Scripts/Enemy/EnemyScriptableObject.cs
@@ -14,7 +14,7 @@ public class EnemyScriptableObject : ScriptableObject
     public float MaxHealth {get => maxHealth; private set => maxHealth = value;}
 
     [SerializeField]
-    float damage;
-    public float Damage {get => damage; private set => damage = value;}
+    int damage;
+    public int Damage {get => damage; private set => damage = value;}
 
 }
